+++ test/packmatic/source_test.exs
@@ -10,13 +10,15 @@ defmodule Packmatic.SourceTest do
   describe "build/1" do
     test "resolves :file" do
       {:ok, file_path} = Briefly.create()
-      {:ok, state} = Packmatic.Source.build({:file, file_path})
+      {:ok, {module, state}} = Packmatic.Source.build({:file, file_path})
+      assert Packmatic.Source.File = module
       assert %Packmatic.Source.File{} = state
     end
 
     test "resolves Packmatic.Source.File" do
       {:ok, file_path} = Briefly.create()
-      {:ok, state} = Packmatic.Source.build({Packmatic.Source.File, file_path})
+      {:ok, {module, state}} = Packmatic.Source.build({Packmatic.Source.File, file_path})
+      assert Packmatic.Source.File = module
       assert %Packmatic.Source.File{} = state
     end
 
@@ -24,7 +26,8 @@ defmodule Packmatic.SourceTest do
       init_arg = :erlang.unique_integer()
       state = %{__struct__: __MODULE__.Source, init_arg: init_arg}
       expect(__MODULE__.Source, :init, fn ^init_arg -> {:ok, state} end)
-      assert {:ok, state} = Packmatic.Source.build({__MODULE__.Source, init_arg})
+      assert {:ok, {module, state}} = Packmatic.Source.build({__MODULE__.Source, init_arg})
+      assert __MODULE__.Source = module
       assert %{__struct__: __MODULE__.Source, init_arg: ^init_arg} = state
     end