+++ CHANGELOG.md
@@ -7,6 +7,22 @@ The format is based on [Keep a Changelog][1], and this project adheres to [Seman
 [1]: https://keepachangelog.com/en/1.0.0/
 [2]: https://semver.org/spec/v2.0.0.html
 
+## [Unreleased][unreleased]
+
+### Added
+
+- Added `Packmatic.Source.Stream`.
+  - Added support for Streams that output IO Lists.
+
+- Updated `Packmatic.Source`.
+  - Added ability for any Source to use any data type as its Source State.
+  - Added ability for any Source to return an updated Source State with new data.
+
+### Changed
+
+- Revised `Packmatic.Source.Dynamic`.
+  - Removed custom resolver; any entry notation now accepted
+
 ## [1.1.0] — 3 October 2020
 
 ### Added